requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
base58==2.1.1
bitarray==1.2.2
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
cytoolz==0.11.2
Deprecated==1.2.13
eth-abi==2.1.1
eth-account==0.5.7
eth-hash==0.3.2
eth-keyfile==0.5.1
eth-keys==0.3.4
eth-rlp==0.3.0
eth-typing==2.3.0
eth-utils==1.10.0
ethereum-dasm==0.1.4
ethereum-etl==2.0.2
evmdasm==0.1.10
frozenlist==1.3.0
hexbytes==0.2.2
idna==3.3
importlib-resources==5.7.1
ipfshttpclient==0.8.0a2
jsonschema==4.5.1
loguru==0.6.0
lru-dict==1.1.7
multiaddr==0.0.9
multidict==6.0.2
netaddr==0.8.0
opensea-api==0.1.7
packaging==21.3
parsimonious==0.8.1
protobuf==3.20.1
pycryptodome==3.14.1
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.2
redis==4.3.3
requests==2.27.1
rlp==2.0.1
six==1.16.0
tabulate==0.8.9
toolz==0.11.2
urllib3==1.26.9
varint==1.0.2
web3==5.29.1
websockets==9.1
wrapt==1.14.1
yapf==0.32.0
yarl==1.7.2
zipp==3.8.0
